Comparison of logistic-SGD
Applying the LogisticRegression and SGDClassifier to the Breast Cancer and Iris Datasets. In the case of LogisticRegression the performance of the algorithm (in terms of accuracy) for different regularization schemes and optimizers while in the case of SGDClassifier, it'll see how different learning rates, loss functions, and the choice of shuffling or not shuffling impacts the performance of the algorithm.